1_Code/requirements.txt
streamlit
pandas
numpy
scikit-learn
torch
torchvision
pillow